# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/monitor/aggregated": {
            "get": {
                "description": "获取指定时间范围内的聚合指标数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "监控"
                ],
                "summary": "获取聚合指标数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "指标类型",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1h",
                        "description": "聚合间隔",
                        "name": "interval",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "开始时间",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/storage.AggregatedMetric"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
...
        "/monitor/latest": {
            "get": {
                "description": "获取指定类型的最新指标数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "监控"
                ],
                "summary": "获取最新指标数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "指标类型",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "限制数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/monitor.MetricRecord"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
...
        "/monitor/range": {
            "get": {
                "description": "获取指定时间范围内的指标数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "监控"
                ],
                "summary": "按时间范围获取指标",
                "parameters": [
                    {
                        "type": "string",
                        "description": "指标类型",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "指标名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "开始时间",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/monitor.MetricRecord"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
...
        "/monitor/system/overview": {
            "get": {
                "description": "获取CPU、内存、磁盘、网络的概览信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "监控"
                ],
                "summary": "获取系统监控概览",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.SystemOverview"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "monitor.MetricRecord": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "labels": {
                    "description": "JSON格式存储标签",
                    "type": "string"
                },
                "metric_name": {
                    "type": "string"
                },
                "metric_type": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "service.CPUOverview": {
            "type": "object",
            "properties": {
                "cores": {
                    "type": "integer"
                },
                "load_avg_1": {
                    "type": "number"
                },
                "load_avg_5": {
                    "type": "number"
                },
                "usage": {
                    "type": "number"
                }
            }
        },
        "service.DiskOverview": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "usage": {
                    "type": "number"
                },
                "used": {
                    "type": "integer"
                }
            }
        },
...
        "service.MemoryOverview": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "usage": {
                    "type": "number"
                },
                "used": {
                    "type": "integer"
                }
            }
        },
        "service.NetworkOverview": {
            "type": "object",
            "properties": {
                "bytes_recv": {
                    "type": "integer"
                },
                "bytes_sent": {
                    "type": "integer"
                }
            }
        },
        "service.SystemOverview": {
            "type": "object",
            "properties": {
                "cpu": {
                    "$ref": "#/definitions/service.CPUOverview"
                },
                "disk": {
                    "$ref": "#/definitions/service.DiskOverview"
                },
                "memory": {
                    "$ref": "#/definitions/service.MemoryOverview"
                },
                "network": {
                    "$ref": "#/definitions/service.NetworkOverview"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "storage.AggregatedMetric": {
            "type": "object",
            "properties": {
                "avg_value": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "max_value": {
                    "type": "number"
                },
                "metric_name": {
                    "type": "string"
                },
                "min_value": {
                    "type": "number"
                },
                "time_bucket": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}